requirements.txt
streamlit
numpy
pandas
scikit-learn
joblib
Pillow
scipy
